Delete Endpoint
Added must be archived before deleting

+++ src/db/users.js
@@ -106,12 +106,23 @@ export const archiveProfile = async (id) => {
 };
 
 export const deleteProfile = async (id) => {
+  const user = await UserModel.findById(id);
+  if (!user.profile_is_archived) {
+    return {
+      status: 400,
+      message: 'Cannot delete a profile that is not archived',
+    };
+  }
   try {
     await deleteAllPosts(id);
     await UserModel.deleteOne({ _id: id });
+    return { status: 200 };
   } catch (error) {
-    console.error('Error deleting profile', error);
-    throw error;
+    console.error('Error deleting profile:', error);
+    return {
+      status: 500,
+      message: 'Error deleting profile',
+    };
   }
 };
